Create `helpers.cpp` and `helpers.h` files
