+++ src-cli/helpers.cpp
@@ -1,0 +1,34 @@
+#include <string>
+
+/// @brief Trim leading and trailing whitespace and quotes from a string
+/// @param str The string to trim
+/// @return The trimmed string
+std::string trim(const std::string &str)
+{
+    size_t start = 0;
+    size_t end = str.size();
+
+    // Trim leading whitespace
+    while (start < end && std::isspace(str[start]))
+    {
+        start++;
+    }
+
+    // Trim trailing whitespace
+    while (end > start && std::isspace(str[end - 1]))
+    {
+        end--;
+    }
+
+    // Trim leading and trailing quotes
+    if (start < end && str[start] == '"')
+    {
+        start++;
+    }
+    if (end > start && str[end - 1] == '"')
+    {
+        end--;
+    }
+
+    return str.substr(start, end - start);
+}
