+++ src-cli/helpers.h
@@ -1,0 +1,11 @@
+#ifndef HELPERS_H
+#define HELPERS_H
+
+#include <string>
+
+/// @brief Trim leading and trailing whitespace and quotes from a string
+/// @param str The string to trim
+/// @return The trimmed string
+std::string trim(const std::string &str);
+
+#endif // HELPERS_H
