+++ src-cli/main.cpp
@@ -7,45 +7,14 @@
 #include <stdexcept>
 #include <sstream>
 
+#include "helpers.h"
+
 // Define a struct to store quotes and authors
 struct Quote
 {
     std::string text;
     std::string author;
 };
-
-/// @brief Trim leading and trailing whitespace and quotes from a string
-/// @param str The string to trim
-/// @return The trimmed string
-std::string trim(const std::string &str)
-{
-    size_t start = 0;
-    size_t end = str.size();
-
-    // Trim leading whitespace
-    while (start < end && std::isspace(str[start]))
-    {
-        start++;
-    }
-
-    // Trim trailing whitespace
-    while (end > start && std::isspace(str[end - 1]))
-    {
-        end--;
-    }
-
-    // Trim leading and trailing quotes
-    if (start < end && str[start] == '"')
-    {
-        start++;
-    }
-    if (end > start && str[end - 1] == '"')
-    {
-        end--;
-    }
-
-    return str.substr(start, end - start);
-}
 
 /// @brief Split a line into fields respecting quoted commas
 /// @param line The CSV line to split
